Remove check that x, y > 0 for mouse move.
This check causes problems on setups with multiple monitors where, e.g., a negative x might be used for a second monitor on the left. The checks shouldn't be needed as long as users input proper amounts to move the cursor by. If one still tries to move the mouse too far, it should just hit the edge of the screen anyway.